backend/requirements.txt
langchain
langchain-openai
langchain_community
flask
flask-cors
faiss-cpu
python-dotenv
pypdf
langgraph-cli